Extract FeedViewController test helper to a separate file
